Changing tests to remove references to selected measures
Several of the tests were referring to the selected measures, which
was used to store the preferences of a user as to what measures they
would see on the dashboard. The code exercised in these tests has
nothing to do with the measures that a user would see on their
dashboard and therefore a plain old user would do.

Addionally, since old factory for users with selected measures can
be removed as where user perferences are stored is changing with the
new Thorax based UI.